refactor: move wagmi code into sdk-react-ui
