+++ packages/sdk-react-ui/src/MetaMaskButton/Balance.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { useNetwork, useAccount, useBalance } from '@metamask/sdk-react';
+import { useNetwork, useAccount, useBalance } from '../hooks/MetaMaskWagmiHooks';
 import { getBalance } from './utils';
 
 const Balance = ({ theme }: { theme: string }) => {
