+++ packages/sdk-react-ui/src/MetaMaskButton/MetaMaskButton.tsx
@@ -1,6 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import Jazzicon, { jsNumberForAddress } from 'react-jazzicon';
-import { useAccount, useConnect, useNetwork, useSDK } from '@metamask/sdk-react';
+import { useAccount, useConnect, useNetwork } from '../hooks/MetaMaskWagmiHooks';
+import { useSDK } from '@metamask/sdk-react';
 import '../style.css';
 import Balance from './Balance';
 import IconNetwork from './IconNetwork';
