+++ packages/sdk-react-ui/src/index.ts
@@ -1,4 +1,6 @@
 import MetaMaskButton from './MetaMaskButton/MetaMaskButton';
+export * from './MetaMaskUIProvider';
+export * from './hooks/MetaMaskWagmiHooks';
 
 // test
 export { MetaMaskButton };
