+++ packages/sdk-react/package.json
@@ -42,14 +42,7 @@
     "test:ci": "echo 'N/A'"
   },
   "dependencies": {
-    "@headlessui/react": "^1.6.6",
-    "@metamask/sdk": "workspace:^",
-    "ethers": "^5.6.9",
-    "mersenne-twister": "^1.1.0",
-    "react-copy-to-clipboard": "^5.1.0",
-    "react-jazzicon": "^1.0.4",
-    "viem": "latest",
-    "wagmi": "^1.3.8"
+    "@metamask/sdk": "workspace:^"
   },
   "devDependencies": {
     "@babel/core": "^7.18.6",
@@ -77,7 +70,6 @@
     "eslint-plugin-jsdoc": "^36.1.0",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-prettier": "^3.4.0",
-    "next": "canary",
     "postcss": "^8.4.12",
     "prettier": "^2.3.0",
     "react": "^18.2.0",
@@ -102,16 +94,8 @@
   "lavamoat": {
     "allowScripts": {
       "@metamask/sdk>eciesjs>secp256k1": false,
-      "ethers>@ethersproject/providers>ws>bufferutil": false,
-      "ethers>@ethersproject/providers>ws>utf-8-validate": false,
-      "wagmi>@coinbase/wallet-sdk>@solana/web3.js>bigint-buffer": false,
-      "wagmi>@coinbase/wallet-sdk>@solana/web3.js>rpc-websockets>utf-8-validate": false,
-      "wagmi>@coinbase/wallet-sdk>keccak": false,
-      "viem>ws>bufferutil": false,
-      "viem>ws>utf-8-validate": false,
-      "wagmi>@wagmi/core>@wagmi/connectors>@coinbase/wallet-sdk>@solana/web3.js>bigint-buffer": false,
-      "wagmi>@wagmi/core>@wagmi/connectors>@coinbase/wallet-sdk>@solana/web3.js>rpc-websockets>utf-8-validate": false,
-      "wagmi>@wagmi/core>@wagmi/connectors>@coinbase/wallet-sdk>keccak": false
+      "@metamask/sdk>react-native>react-devtools-core>ws>bufferutil": false,
+      "@metamask/sdk>react-native>react-devtools-core>ws>utf-8-validate": false
     }
   }
 }
